requirements.txt
# =============================================================================
# AI Agent + Quant Trading System — Python 3.11+
# Install: pip install -r requirements.txt
# =============================================================================

# --- Strands Agents SDK ------------------------------------------------------
strands-agents>=1.29.0

# --- AWS SDK (Bedrock backend) -----------------------------------------------
boto3>=1.42.0
botocore>=1.42.0

# --- Broker: Alpaca Markets --------------------------------------------------
alpaca-py>=0.43.0

# --- Market Data -------------------------------------------------------------
alpha-vantage>=2.3.0
yfinance>=0.2.40
lxml>=5.0.0
requests>=2.32.0

# --- Quantitative / Data Science ---------------------------------------------
pandas>=2.2.0
numpy>=1.26.0
ta>=0.11.0
scipy>=1.13.0

# --- Configuration -----------------------------------------------------------
pydantic>=2.7.0
pydantic-settings>=2.3.0

# --- Scheduling --------------------------------------------------------------
APScheduler>=3.10.0
pytz>=2024.1

# --- Utilities ---------------------------------------------------------------
python-dotenv>=1.0.0
python-dateutil>=2.9.0
tenacity>=8.3.0
PyYAML>=6.0.1

# --- Cloud / AgentCore Runtime -----------------------------------------------
bedrock-agentcore>=0.0.5
fastapi>=0.135.0
uvicorn>=0.30.0

# --- Logging -----------------------------------------------------------------
python-json-logger>=2.0.7
